Tests restructurization WIP
