--- test/Enecuum/Framework/Testing/Runtime.hs
@@ -1,6 +1,0 @@
-module Enecuum.Framework.Testing.Runtime
-  ( module X
-  ) where
-
-import Enecuum.Framework.Testing.Environment.TestRuntime     as X
-import Enecuum.Framework.Testing.Node.Runtime                as X
