--- test/Enecuum/FunctionalTesting.hs
@@ -1,9 +1,0 @@
-module Enecuum.FunctionalTesting
-  ( module X
-  ) where
-
-import           Enecuum.FunctionalTesting.Core.LoggerRuntime as X
-import           Enecuum.FunctionalTesting.Core.Interpreters as X
-import           Enecuum.FunctionalTesting.Framework.NodeRuntime as X
-import           Enecuum.FunctionalTesting.Framework.Interpreters as X
-import           Enecuum.FunctionalTesting.TestRuntime as X
