--- test/Enecuum/FunctionalTesting/Core/Interpreters.hs
@@ -1,6 +1,0 @@
-module Enecuum.FunctionalTesting.Core.Interpreters
-  ( module X
-  ) where
-
-import           Enecuum.FunctionalTesting.Core.Interpreters.Logger as X
-import           Enecuum.FunctionalTesting.Core.Interpreters.CoreEffect as X
