--- test/Enecuum/FunctionalTesting/Core/LoggerRuntime.hs
@@ -1,8 +1,0 @@
-module Enecuum.FunctionalTesting.Core.LoggerRuntime where
-
-import           Enecuum.Prelude
-
-import qualified Enecuum.FunctionalTesting.Types as T
-
-createLoggerRuntime :: IO T.LoggerRuntime
-createLoggerRuntime = T.LoggerRuntime <$> newTVarIO []
