+++ test/Enecuum/Testing/Framework/Interpreters/Node.hs
@@ -1,0 +1,30 @@
+module Enecuum.Testing.Framework.Interpreters.Node where
+
+import Enecuum.Prelude
+
+import qualified Enecuum.Language                                  as L
+
+import qualified Enecuum.Testing.RLens                             as RLens
+import qualified Enecuum.Testing.Types                             as T
+import qualified Enecuum.Testing.Core.Interpreters                 as Impl
+import qualified Enecuum.Testing.Framework.Interpreters.Networking as Impl
+import qualified Enecuum.Testing.Framework.Interpreters.State      as Impl
+
+-- | Interpret NodeL.
+interpretNodeL :: T.NodeRuntime -> L.NodeF a -> IO a
+
+interpretNodeL nodeRt (L.EvalStateAtomically statefulAction next) = do
+  next <$> (atomically $ Impl.runStateL nodeRt statefulAction)
+
+interpretNodeL nodeRt (L.EvalGraphIO (L.GraphAction _ ioRunner act) next) = do
+  next <$> ioRunner act
+
+interpretNodeL nodeRt (L.EvalNetworking networkingAction next) =
+  next <$> Impl.runNetworkingL nodeRt networkingAction
+
+interpretNodeL nodeRt (L.EvalCoreEffectNodeF coreEffect next) =
+  next <$> Impl.runCoreEffect (nodeRt ^. RLens.loggerRuntime) coreEffect
+
+-- | Runs node language.
+runNodeL :: T.NodeRuntime -> L.NodeL a -> IO a
+runNodeL nodeRt = foldFree (interpretNodeL nodeRt)
